+++ Source/Krypton Components/Krypton.Navigator/Dragging/DropSolidWindow.cs
@@ -1,4 +1,4 @@
-﻿#region BSD License
+#region BSD License
 /*
  * 
  * Original BSD 3-Clause License (https://github.com/ComponentFactory/Krypton/blob/master/LICENSE)
@@ -81,11 +81,10 @@ public class DropSolidWindow : KryptonForm
             {
                 _solidRect = value;
 
-                var area = Screen.GetWorkingArea(this);
-
-                var bounds = new Rectangle(value.Location - (Size)area.Location, value.Size);
-
-                DesktopBounds = bounds;
+                // DrawRect/SolidRect are in screen coordinates; DesktopBounds expects screen coordinates.
+                // Do not subtract working-area origin (fixes #2935: border on wrong monitor when
+                // drop indicator is on a secondary monitor).
+                DesktopBounds = value;
 
                 Refresh();
             }